+++ CHANGES.md
@@ -69,6 +69,7 @@ UI:
 -   Adjusted tooltip style & layout
 -   Overall page layout adjustment for `Details and Content` page
 -   fixed: keywords & themes extraction might produce duplicate keywords for different files
+-   Restricted dataset contact point display options to team members, team or org and made it save in the registry.
 -   Added ability to set the owning org unit in Add Dataset flow.
 -   Stopped keywords & themes extraction producing duplicate keywords for different files
 -   Capped the maximum input to retext for keyword extraction to prevent browser freezeup